backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.9.0
torchvision==0.24.0
Pillow==11.0.0
pydantic
python-dotenv==1.0.0
google-generativeai